Compile kotlin-runtime-minimal.jar, include into bootstrap compiler
Include minimal runtime into bootstrap compiler (step 1 of 2 in the build)
instead of the full runtime because full runtime intersects with compiler on
classes from core/ (descriptors, JDR, ...) and these classes are in fact of
different versions. This effectively means that reflection can't be used from
the compiler, otherwise bootstrap compiler will not link (in fact it will only
fail at runtime when/if the reflection usage is reached).